Add initial Jupyter notebook for Swarm experimentation
This commit introduces a new Jupyter notebook (1.ipynb) that outlines the rationale for using Swarm, a lightweight and scalable framework. The notebook includes installation commands for the Swarm library and dotenv package, along with initial code to set up the environment. An error related to the missing dotenv module is also documented in the notebook.